Add shorthand functions and parameters, and map/1
The map/1 and map_values/1 functions are equivalent to
    def map(f): [.[] | f]
and
    def map_values(f): .[] |= f
respectively. This commit introduces named parameters like f above and
uses them to define these two functions with the bodies above, compiled
from a string.

An argument to a function is bound to the corresponding parameter as a
filter, a named zero-parameter function whose evaluation is the filter
passed in.