Use instanceof for Array type checks
Previously, jqjs very conservatively checked for the Array constructor
explicitly, which was problematic for custom subtypes. Now, it uses
instanceof to check the entire prototype chain; this will cause a
behavioural change for any values extending the Array type that were
previously treated as object values, but it seems unlikely that this
creates a real problem.